# Conflicts:
#	pom.xml
#	src/main/java/com/pax/market/api/sdk/java/api/BaseThirdPartySysApi.java
#	src/main/java/com/pax/market/api/sdk/java/api/client/ThirdPartySysApiClient.java

+++ src/main/java/com/pax/market/api/sdk/java/api/util/ThirdPartySysHttpUtils.java
@@ -80,8 +80,9 @@ public abstract class ThirdPartySysHttpUtils {
      * @param headerMap      the header map
      * @return the string
      */
-    public static String request(String requestUrl, String requestMethod, int connectTimeout, int readTimeout, String userData, Map<String, String> headerMap){
-		return request(requestUrl, requestMethod, connectTimeout, readTimeout, userData, false, headerMap, null);
+    public static String request(String requestUrl, String requestMethod, int connectTimeout, int readTimeout, String userData, Map<String, String> headerMap, int retryTimes){
+//		return request(requestUrl, requestMethod, connectTimeout, readTimeout, userData, false, headerMap, null);
+		return request(requestUrl, requestMethod, connectTimeout, readTimeout, userData, headerMap, null, retryTimes);
 	}
 
     /**
@@ -95,8 +96,8 @@ public abstract class ThirdPartySysHttpUtils {
      * @param headerMap      the header map
      * @return the string
      */
-    public static String compressRequest(String requestUrl, String requestMethod, int connectTimeout, int readTimeout, String userData, Map<String, String> headerMap){
-		return request(requestUrl, requestMethod, connectTimeout, readTimeout, userData, true, headerMap, null);
+    public static String compressRequest(String requestUrl, String requestMethod, int connectTimeout, int readTimeout, String userData, Map<String, String> headerMap, int retryTimes){
+		return compressRequest(requestUrl, requestMethod, connectTimeout, readTimeout, userData, headerMap, null, retryTimes);
 	}
 
     /**
@@ -111,8 +112,22 @@ public abstract class ThirdPartySysHttpUtils {
      * @param saveFilePath   the save file path
      * @return the string
      */
-    public static String request(String requestUrl, String requestMethod, int connectTimeout, int readTimeout, String userData, Map<String, String> headerMap, String saveFilePath){
-		return request(requestUrl, requestMethod, connectTimeout, readTimeout, userData, false, headerMap, saveFilePath);
+    public static String request(String requestUrl, String requestMethod, int connectTimeout, int readTimeout, String userData, Map<String, String> headerMap, String saveFilePath, int retryTimes){
+//		return request(requestUrl, requestMethod, connectTimeout, readTimeout, userData, false, headerMap, saveFilePath);
+		try {
+			return RetryUtils.retry(() -> {
+				return request(requestUrl, requestMethod, connectTimeout, readTimeout, userData, false, headerMap, saveFilePath);
+			}, e -> isExceptionShouldRetry(e), retryTimes);
+		} catch (Exception e) {
+			FileUtils.deleteFile(saveFilePath);
+			logger.error("Exception Occurred. Details: {}", e.toString());
+			if(e instanceof IOException) {
+				return EnhancedJsonUtils.getSdkJson(ResultCode.SDK_UN_CONNECT);
+			}else{
+				return EnhancedJsonUtils.getSdkJson(ResultCode.SDK_RQUEST_EXCEPTION);
+			}
+
+		}
 	}
 
     /**
@@ -127,17 +142,42 @@ public abstract class ThirdPartySysHttpUtils {
      * @param saveFilePath   the save file path
      * @return the string
      */
-    public static String compressRequest(String requestUrl, String requestMethod, int connectTimeout, int readTimeout, String userData, Map<String, String> headerMap, String saveFilePath){
-		return request(requestUrl, requestMethod, connectTimeout, readTimeout, userData, true, headerMap, saveFilePath);
+    public static String compressRequest(String requestUrl, String requestMethod, int connectTimeout, int readTimeout, String userData, Map<String, String> headerMap, String saveFilePath, int retryTimes){
+		try {
+			return RetryUtils.retry(() -> {
+				return request(requestUrl, requestMethod, connectTimeout, readTimeout, userData, true, headerMap, saveFilePath);
+			}, e -> isExceptionShouldRetry(e), retryTimes);
+		} catch (Exception e) {
+			if(StringUtils.isNotBlank(saveFilePath)) {
+				FileUtils.deleteFile(saveFilePath);
+			}
+			logger.error("Occurred. Details: {}", e.toString());
+			return EnhancedJsonUtils.getSdkJson(ResultCode.SDK_UN_CONNECT);
+		}
+	}
+
+	private static boolean isExceptionShouldRetry(Throwable e) {
+    	if(e instanceof ConnectException){
+    		return true;
+		} else if(e instanceof SocketTimeoutException){
+    		return true;
+		}else{
+    		return false;
+		}
 	}
 
 	private static String request(String requestUrl, String requestMethod, int connectTimeout, int readTimeout, String userData, boolean compressData,
-								  Map<String, String> headerMap, String saveFilePath) {
+								  Map<String, String> headerMap, String saveFilePath) throws ConnectException, SocketTimeoutException{
 		HttpURLConnection urlConnection = null;
 		try {
 			urlConnection = getConnection(requestUrl, connectTimeout, readTimeout);
 			return finalRequest(urlConnection, requestMethod, userData, compressData, headerMap, saveFilePath);
 		} catch (IOException e) {
+			if(e instanceof ConnectException){
+				throw (ConnectException)e;
+			}else if(e instanceof SocketTimeoutException){
+				throw (SocketTimeoutException)e;
+			}
 			logger.error("IOException Occurred. Details: {}", e.toString());
 		} finally {
 			if(urlConnection != null) {
@@ -148,7 +188,7 @@ public abstract class ThirdPartySysHttpUtils {
 	}
 
 	private static String finalRequest(HttpURLConnection urlConnection, String requestMethod, String userData, boolean compressData,
-									   Map<String, String> headerMap, String saveFilePath) {
+									   Map<String, String> headerMap, String saveFilePath) throws ConnectException,SocketTimeoutException{
 		StringBuilder stringBuilder = new StringBuilder();
 		BufferedReader bufferedReader = null;
 		FileOutputStream fileOutputStream = null;
@@ -255,13 +295,20 @@ public abstract class ThirdPartySysHttpUtils {
 //				
 //			}
 		} catch (SocketTimeoutException localSocketTimeoutException) {
-			FileUtils.deleteFile(filePath);
-			logger.error("SocketTimeoutException Occurred. Details: {}", localSocketTimeoutException.toString());
-			return EnhancedJsonUtils.getSdkJson(ResultCode.SDK_CONNECT_TIMEOUT);
+			if(StringUtils.containsIgnoreCase(localSocketTimeoutException.toString(),"Read timed out")){
+				FileUtils.deleteFile(filePath);
+				logger.error("SocketTimeoutException Occurred. Details: {}", localSocketTimeoutException.toString());
+				return EnhancedJsonUtils.getSdkJson(ResultCode.SDK_CONNECT_TIMEOUT);
+			}else{
+				throw localSocketTimeoutException;
+			}
+
 		} catch (ConnectException localConnectException) {
-			FileUtils.deleteFile(filePath);
-			logger.error("ConnectException Occurred. Details: {}", localConnectException.toString());
-			return EnhancedJsonUtils.getSdkJson(ResultCode.SDK_UN_CONNECT);
+//			FileUtils.deleteFile(filePath);
+//			logger.error("ConnectException Occurred. Details: {}", localConnectException.toString());
+//			return EnhancedJsonUtils.getSdkJson(ResultCode.SDK_UN_CONNECT);
+			throw localConnectException;
+
 		} catch (FileNotFoundException fileNotFoundException) {
 			FileUtils.deleteFile(filePath);
 			logger.error("FileNotFoundException Occurred. Details: {}", fileNotFoundException.toString());
